fix: focusing animation of message search results has been modified to apply only to the message component
