+++ packages/uikit-react-native/src/fragments/createGroupChannelFragment.tsx
@@ -1,5 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
-import { Animated, Easing } from 'react-native';
+import React, { useCallback, useMemo, useState } from 'react';
 
 import { useGroupChannelMessages } from '@sendbird/uikit-chat-hooks';
 import {
@@ -17,7 +16,6 @@ import GroupChannelMessageRenderer from '../components/GroupChannelMessageRender
 import NewMessagesButton from '../components/NewMessagesButton';
 import ScrollToBottomButton from '../components/ScrollToBottomButton';
 import StatusComposition from '../components/StatusComposition';
-import { MESSAGE_FOCUS_ANIMATION_DELAY, MESSAGE_SEARCH_SAFE_SCROLL_DELAY } from '../constants';
 import createGroupChannelModule from '../domain/groupChannel/module/createGroupChannelModule';
 import type {
   GroupChannelFragment,
@@ -91,13 +89,9 @@ const createGroupChannelFragment = (initModule?: Partial<GroupChannelModule>): G
       startingPoint: internalSearchItem?.startingPoint,
     });
 
-    const MessageComponent: GroupChannelProps['MessageList']['renderMessage'] = useCallback(
-      withFocusingAnimation(renderMessage ? (props) => <>{renderMessage(props)}</> : GroupChannelMessageRenderer),
-      [renderMessage],
-    );
-
-    const _renderMessage: GroupChannelProps['MessageList']['renderMessage'] = useFreshCallback((props) => {
-      return <MessageComponent {...props} />;
+    const renderItem: GroupChannelProps['MessageList']['renderMessage'] = useFreshCallback((props) => {
+      if (renderMessage) return renderMessage(props);
+      return <GroupChannelMessageRenderer {...props} />;
     });
 
     const memoizedFlatListProps = useMemo(
@@ -211,7 +205,7 @@ const createGroupChannelFragment = (initModule?: Partial<GroupChannelModule>): G
             onResetMessageList={onResetMessageList}
             enableMessageGrouping={enableMessageGrouping}
             currentUserId={currentUser?.userId}
-            renderMessage={_renderMessage}
+            renderMessage={renderItem}
             messages={messages}
             newMessages={newMessages}
             onTopReached={prev}
@@ -253,35 +247,6 @@ function shouldRenderInput(channel: SendbirdGroupChannel) {
   }
 
   return true;
-}
-
-function withFocusingAnimation<P extends unknown & { focused: boolean }>(Component: React.ComponentType<P>) {
-  return React.memo<P>((props) => {
-    const translateY = useRef(new Animated.Value(0)).current;
-
-    useEffect(() => {
-      if (props.focused) {
-        setTimeout(() => {
-          Animated.sequence(
-            [
-              { toValue: -10, duration: 500 },
-              { toValue: 0, duration: 100 },
-              { toValue: -10, duration: 200 },
-              { toValue: 0, duration: 100 },
-            ].map((value) =>
-              Animated.timing(translateY, { ...value, useNativeDriver: true, easing: Easing.inOut(Easing.ease) }),
-            ),
-          ).start();
-        }, MESSAGE_SEARCH_SAFE_SCROLL_DELAY + MESSAGE_FOCUS_ANIMATION_DELAY);
-      }
-    }, [props.focused]);
-
-    return (
-      <Animated.View style={{ transform: [{ translateY }] }}>
-        <Component {...props} />
-      </Animated.View>
-    );
-  });
 }
 
 export default createGroupChannelFragment;
